Add loader to landing page, improve no results message
This adds a loader to the landing page on serach and also
styles the no-results message more nicely.

It also removes the main module in favor of just one big
app module. My current understanding is that there should not
have been separate modules to begin with since this is just
one small and simple app with a shared scope.